fix(wasm32): fence the post-beta unix/libc/usize regressions out of the wasm build
Found by running wasm/wasm-build.sh on the v0.3 tree after the pgcrypto fence — none of these had
a wasm arm; nothing in CI builds wasm32, so they accumulated since the v0.3-beta (2026-08-04):
- pg_path set_pglocale_pgservice: libc::setlocale/LC_ALL (no such thing on wasi) — non-wasm only
- fd copydir clone_file: neither HAVE_COPYFILE nor HAVE_COPY_FILE_RANGE arm existed — ereport
  ERRCODE_FEATURE_NOT_SUPPORTED like C's #else would have to
- pgrc2_batch StrCell::image_ptr: usize::from_ne_bytes on the 8-byte tail — read as u64
- pgclient unix_connect: the AF_UNIX arm lacked its cfg(not(wasm)) (duplicate definition on wasm)
- injection_point GATE: 1 << 32 in a usize — u64 on every pointer width
- hba RADIUS resolution: libc::SOCK_DGRAM missing on wasi — POSIX value carried
- sqe bankstats: std::os::unix::fs::FileExt — wasi FileExt behind wasi_ext (pgrcolumnar precedent)
- uuid_ossp get_node_id: getifaddrs/sockaddr missing on wasi — no MAC, random node id
Native code paths are unchanged.

+++ crates/backend/access/pgrcolumnar2/pgrc2_batch/src/strview.rs
@@ -240,7 +240,9 @@ impl StrCell {
     #[inline]
     pub fn image_ptr(&self) -> *const u8 {
         debug_assert!(!self.is_inline());
-        usize::from_ne_bytes(self.tail) as *const u8
+        // tail is always 8 bytes (line 141 packs the pointer as u64), so read it
+        // back as u64 on every pointer width (wasm32 has a 4-byte usize).
+        u64::from_ne_bytes(self.tail) as usize as *const u8
     }
 
     /// The string bytes.
